Add Bazel BUILD file.
The BUILD file trivially wraps the bazel/openssl.bzl rule's file, which is
generated via the following command:
    cd bazel
    ./log_to_bzl.py build.log.x86_64 --format bazel > openssl.bzl

The build.log.x86_64 file is generated using the following commands, kept in
source for the record:
    ./config no-shared no-dso
    make -j8 buld_libs > build.log.x86_64

Note that the 'external' directory conflicts with Bazel's 'external' directory.
Move 'external' to 'third_party'.  The only affected components are the perl
modules, i.e., *.pm.  Use equivalents to the following in the BUILD file to
accommodate to this change:
    export PERL5LIB=third_party/perl

It is worth noting that the configdata.pm file is generated by running ./config.
This has been fully bazelised but is only tested on Ubuntu 16 x86_64.  It might
very well break on other platforms.