Add openssl.sierra.bzl.
Restructured the bazel/ directory to add more support for building on sierra.

Can still only build on xenial.  But you can make it build on sierra by changing
the load() function in the BUILD file to load the 'openssl.sierra.bzl' file.

TODO: Formalize the building for sierra and xenial using select() and
config_setting().